% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{range}
\alias{range}
\title{Find the range of a vector}
\usage{
range(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
A vector containing the resulting value when the min is subtracted from the max.
}
\description{
Find the range of a vector
}
\examples{
range(1:10)
}
